Separate freq module
