+++ tooling/ast_fuzzer/src/program/mod.rs
@@ -17,6 +17,7 @@ use types::type_depth;
 use crate::Config;
 
 mod expr;
+pub(crate) mod freq;
 mod func;
 mod types;
